refactor: move exception class to main package
