+++ deeplearning/src/main/java/fr/cyu/smartread/deeplearning/IncompatibleShapeException.java
@@ -1,4 +1,4 @@
-package fr.cyu.smartread.deeplearning.losses;
+package fr.cyu.smartread.deeplearning;
 
 import org.ejml.data.DMatrixRMaj;
