+++ deeplearning/src/main/java/fr/cyu/smartread/deeplearning/losses/AbstractLoss.java
@@ -1,5 +1,6 @@
 package fr.cyu.smartread.deeplearning.losses;
 
+import fr.cyu.smartread.deeplearning.IncompatibleShapeException;
 import fr.cyu.smartread.deeplearning.UtilityOperationsMatrix;
 import fr.cyu.smartread.deeplearning.activations.NoTrainingComputationsPerformedException;
 import org.ejml.data.DMatrixRMaj;
