+++ deeplearning/src/test/java/fr/cyu/smartread/deeplearning/losses/AbstractLossTest.java
@@ -1,5 +1,6 @@
 package fr.cyu.smartread.deeplearning.losses;
 
+import fr.cyu.smartread.deeplearning.IncompatibleShapeException;
 import fr.cyu.smartread.deeplearning.activations.NoTrainingComputationsPerformedException;
 import org.ejml.data.DMatrixRMaj;
 import org.junit.jupiter.api.Assertions;
